Chore: Rename files and directories
