+++ scripts/test-vectors/index-20220613.ts
@@ -1,3 +1,6 @@
+// Date represents the date of the issue with the test vectors sent to Zondax:
+// https://github.com/Zondax/ledger-icp/issues/161
+// We'll keep each script for reproducibility.
 import { IDL } from "@dfinity/candid";
 import { Principal } from "@dfinity/principal";
 import { NeuronId as PbNeuronId } from "../../proto/base_types_pb";
@@ -30,7 +33,7 @@ import {
   fromStartDissolvingRequest,
   fromStopDissolvingRequest,
 } from "../../src/canisters/governance/request.proto.converters";
-import { ListNeuronsFn, ManageNeuronFn } from "./governanceCandid";
+import { ListNeuronsFn, ManageNeuronFn } from "./governance.idl";
 import { createBlob, writeToJson } from "./utils";
 
 const mockNeuronId = BigInt(15374508381553347371);
