+++ scripts/test-vectors/utils.ts
